+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryEffectHandlerTest.kt
@@ -934,7 +934,11 @@ class PatientSummaryEffectHandlerTest {
             dateOfBirth = null,
         )
     )
-    val medicalHistory = TestData.medicalHistory(isSmoking = Yes)
+    val medicalHistory = TestData.medicalHistory(
+        isSmoking = Yes,
+        hasDiabetes = Yes,
+        cholesterol = 400f,
+    )
     val bloodPressure = TestData.bloodPressureMeasurement(
         UUID.fromString("3e8c246f-91b9-4f8c-81fe-91b67ac0a2d5"),
         systolic = 130,
@@ -948,14 +952,13 @@ class PatientSummaryEffectHandlerTest {
         patientUuid = patientUuid,
         defaultHistoryUuid = uuidGenerator.v4()
     )) doReturn medicalHistory
-    whenever(cvdRiskRepository.getCVDRiskImmediate(patientUuid)) doReturn
-        cvdRisk
+    whenever(cvdRiskRepository.getCVDRiskImmediate(patientUuid)) doReturn cvdRisk
 
     //when
     testCase.dispatch(CalculateLabBasedCVDRisk(patient = patient))
 
     //then
-    testCase.assertOutgoingEvents(CVDRiskCalculated(cvdRisk, CVDRiskRange(7, 14)))
+    testCase.assertOutgoingEvents(CVDRiskCalculated(cvdRisk, CVDRiskRange(14, 14)))
   }
 
   @Test
@@ -983,11 +986,21 @@ class PatientSummaryEffectHandlerTest {
   fun `when load statin info effect is received, then load statin info`() {
     //given
     val bmiReading = BMIReading(height = 177f, weight = 53f)
+
+    whenever(patientRepository.patientImmediate(patientUuid)) doReturn TestData.patient(
+        uuid = patientUuid,
+        patientAgeDetails = PatientAgeDetails(
+            ageValue = 55,
+            ageUpdatedAt = Instant.parse("2018-01-01T00:00:00Z"),
+            dateOfBirth = null,
+        )
+    )
+    val medicalHistory = TestData.medicalHistory(isSmoking = Yes)
+
     whenever(medicalHistoryRepository.historyForPatientOrDefaultImmediate(
         defaultHistoryUuid = uuidGenerator.v4(),
         patientUuid = patientUuid
-    )) doReturn
-        TestData.medicalHistory(isSmoking = Yes)
+    )) doReturn medicalHistory
 
     whenever(patientAttributeRepository.getPatientAttributeImmediate(patientUuid)) doReturn
         TestData.patientAttribute(reading = bmiReading)
@@ -999,12 +1012,12 @@ class PatientSummaryEffectHandlerTest {
     testCase.dispatch(LoadStatinInfo(patientUuid))
 
     //then
-    testCase.assertOutgoingEvents(StatinInfoLoaded(StatinInfo(
-        canPrescribeStatin = true,
-        cvdRisk = CVDRiskRange(27, 27),
-        isSmoker = Yes,
-        bmiReading = bmiReading
-    )))
+    testCase.assertOutgoingEvents(StatinInfoLoaded(
+        age = 55,
+        medicalHistory = medicalHistory,
+        riskRange = CVDRiskRange(27, 27),
+        bmiReading = bmiReading,
+    ))
   }
 
   @Test
@@ -1028,6 +1041,17 @@ class PatientSummaryEffectHandlerTest {
     testCase.assertNoOutgoingEvents()
 
     verify(uiActions).openBMIEntrySheet(patientUuid)
+    verifyNoMoreInteractions(uiActions)
+  }
+
+  @Test
+  fun `when open cholesterol entry sheet effect is received, then open cholesterol entry sheet`() {
+    // when
+    testCase.dispatch(OpenCholesterolEntrySheet(patientUuid = patientUuid))
+
+    // then
+    testCase.assertNoOutgoingEvents()
+    verify(uiActions).openCholesterolEntrySheet(patientUuid = patientUuid)
     verifyNoMoreInteractions(uiActions)
   }
 }